Fix test compilation

+++ crates/augmented/data/augmented-playhead/src/lib.rs
@@ -114,7 +114,6 @@ impl PlayHead {
 #[cfg(test)]
 mod test {
     use crate::{PlayHead, PlayHeadOptions};
-    use audio_processor_testing_helpers::assert_f_eq;
 
     #[test]
     fn test_accept_samples() {
@@ -163,10 +162,10 @@ mod test {
         let mut play_head = PlayHead::new(options);
         play_head.accept_samples(sample_count);
         assert!((play_head.position_seconds() - 128.0).abs() < f32::EPSILON);
-        assert!((play_head.position_beats() - 256.0).abs() < f32::EPSILON);
+        assert!((play_head.position_beats() - 256.0).abs() < f64::EPSILON);
         play_head.accept_samples(sample_count / 2);
         assert!((play_head.position_seconds() - 192.0).abs() < f32::EPSILON);
-        assert!((play_head.position_beats() - 384.0).abs() < f32::EPSILON);
+        assert!((play_head.position_beats() - 384.0).abs() < f64::EPSILON);
     }
 
     // #[test]
